Refactor and enhance option handling and CFG logic
- Improved type safety with `static_cast` in `language.hpp`.
- Introduced `OptionCollection` for centralized option management.
- Enhanced control flow graph with `SplitNode` and replacement methods.
- Refactored `FunctionInliner` and `ILOptimizer` to use `OptionCollection`.
- Updated `MurmurHash3` to use `size_t` for length parameters.
- Added utility methods for list manipulation in `ilist.hpp`.
- Integrated `OptionCollection` into `Compiler` for better configuration.
- Updated submodule reference in `zstd` to reflect changes.

+++ backend/include/core/language.hpp
@@ -29,7 +29,7 @@ namespace HXSL
 
 	enum ParameterFlags : uint8_t
 	{
-		ParameterFlags_None = -1,
+		ParameterFlags_None = static_cast<uint8_t>(-1),
 		ParameterFlags_In = 0,
 		ParameterFlags_Out = 1,
 		ParameterFlags_InOut = 2,
